requirements.txt
Django>=5.2
djangorestframework
pillow
weasyprint
django-cors-headers
reportlab
